requirements.txt
google-generativeai>=0.8.3
vaderSentiment>=3.3.2
websockets>=12.0
pyttsx3>=2.90
python-dotenv>=1.0.1
edge-tts>=6.1.9
azure-cognitiveservices-speech>=1.36.0